refactor: extract topic creation
